# openapi_client/models/entry_information.py
# coding: utf-8

"""
    FastAPI

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 0.1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json

from pydantic import BaseModel, ConfigDict, StrictBool, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List
from typing import Optional, Set
from typing_extensions import Self

class EntryInformation(BaseModel):
    """
    EntryInformation
    """ # noqa: E501
    id: StrictInt
    namespace: StrictStr
    name: StrictStr
    version: StrictStr
    category: StrictStr
    description: StrictStr
    details: Dict[str, Any]
    tags: List[StrictStr]
    num_stars: StrictInt
    starred_by_point_of_view: StrictBool
    __properties: ClassVar[List[str]] = ["id", "namespace", "name", "version", "category", "description", "details", "tags", "num_stars", "starred_by_point_of_view"]

    model_config = ConfigDict(
        populate_by_name=True,
        validate_assignment=True,
        protected_namespaces=(),
    )


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Optional[Self]:
        """Create an instance of EntryInformation from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        excluded_fields: Set[str] = set([
        ])

        _dict = self.model_dump(
            by_alias=True,
            exclude=excluded_fields,
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
        """Create an instance of EntryInformation from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "id": obj.get("id"),
            "namespace": obj.get("namespace"),
            "name": obj.get("name"),
            "version": obj.get("version"),
            "category": obj.get("category"),
            "description": obj.get("description"),
            "details": obj.get("details"),
            "tags": obj.get("tags"),
            "num_stars": obj.get("num_stars"),
            "starred_by_point_of_view": obj.get("starred_by_point_of_view")
        })
        return _obj