# scripts/near-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/registry/upload_file": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_v1_registry_upload_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_v1_registry_upload_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/registry/download_file": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Download File",
        "operationId": "download_file_v1_registry_download_file_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_download_file_v1_registry_download_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/registry/upload_metadata": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Upload Metadata",
        "operationId": "upload_metadata_v1_registry_upload_metadata_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_metadata_v1_registry_upload_metadata_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/registry/download_metadata": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Download Metadata",
        "operationId": "download_metadata_v1_registry_download_metadata_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_download_metadata_v1_registry_download_metadata_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/registry/list_files": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "List Files",
        "description": "List all files that belong to a entry.",
        "operationId": "list_files_v1_registry_list_files_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_list_files_v1_registry_list_files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Filename"
                  },
                  "type": "array",
                  "title": "Response List Files V1 Registry List Files Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/registry/list_entries": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "List Entries",
        "operationId": "list_entries_v1_registry_list_entries_post",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Namespace"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Category"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Tags"
            }
          },
          {
            "name": "total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 32,
              "title": "Total"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "show_hidden",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Show Hidden"
            }
          },
          {
            "name": "show_latest_version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Show Latest Version"
            }
          },
          {
            "name": "starred_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Starred By"
            }
          },
          {
            "name": "star_point_of_view",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Star Point Of View"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntryInformation"
                  },
                  "title": "Response List Entries V1 Registry List Entries Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_download_file_v1_registry_download_file_post": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "entry_location": {
            "$ref": "#/components/schemas/EntryLocation"
          }
        },
        "type": "object",
        "required": [
          "path",
          "entry_location"
        ],
        "title": "Body_download_file_v1_registry_download_file_post"
      },
      "Body_download_metadata_v1_registry_download_metadata_post": {
        "properties": {
          "entry_location": {
            "$ref": "#/components/schemas/EntryLocation"
          }
        },
        "type": "object",
        "required": [
          "entry_location"
        ],
        "title": "Body_download_metadata_v1_registry_download_metadata_post"
      },
      "Body_list_files_v1_registry_list_files_post": {
        "properties": {
          "entry_location": {
            "$ref": "#/components/schemas/EntryLocation"
          }
        },
        "type": "object",
        "required": [
          "entry_location"
        ],
        "title": "Body_list_files_v1_registry_list_files_post"
      },
...
      "Body_upload_file_v1_registry_upload_file_post": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "path",
          "file",
          "namespace",
          "name",
          "version"
        ],
        "title": "Body_upload_file_v1_registry_upload_file_post"
      },
      "Body_upload_metadata_v1_registry_upload_metadata_post": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/EntryMetadataInput"
          },
          "entry_location": {
            "$ref": "#/components/schemas/EntryLocation"
          }
        },
        "type": "object",
        "required": [
          "metadata",
          "entry_location"
        ],
        "title": "Body_upload_metadata_v1_registry_upload_metadata_post"
      },
...
      "EntryInformation": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "details": {
            "type": "object",
            "title": "Details"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "num_stars": {
            "type": "integer",
            "title": "Num Stars"
          },
          "starred_by_point_of_view": {
            "type": "boolean",
            "title": "Starred By Point Of View"
          }
        },
        "type": "object",
        "required": [
          "id",
          "namespace",
          "name",
          "version",
          "category",
          "description",
          "details",
          "tags",
          "num_stars",
          "starred_by_point_of_view"
        ],
        "title": "EntryInformation"
      },
      "EntryLocation": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "name",
          "version"
        ],
        "title": "EntryLocation"
      },
      "EntryMetadata": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "details": {
            "type": "object",
            "title": "Details"
          },
          "show_entry": {
            "type": "boolean",
            "title": "Show Entry"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "category",
          "description",
          "tags",
          "details",
          "show_entry",
          "name",
          "version"
        ],
        "title": "EntryMetadata"
      },
      "EntryMetadataInput": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "details": {
            "type": "object",
            "title": "Details"
          },
          "show_entry": {
            "type": "boolean",
            "title": "Show Entry"
          }
        },
        "type": "object",
        "required": [
          "category",
          "description",
          "tags",
          "details",
          "show_entry"
        ],
        "title": "EntryMetadataInput"
      },
      "Filename": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "Filename"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}